refactor: Organization
