+++ exercise/C#/day20/Reindeer.Web.Tests/Common/ReindeerApiContract.cs
@@ -1,0 +1,3 @@
+namespace Reindeer.Web.Tests.Common;
+
+public record ReindeerApiContract(HttpRequestMessage? Request, HttpResponseMessage Response);
