+++ exercise/C#/day20/Reindeer.Web.Tests/Common/ReindeerWebApplicationFactory.cs
@@ -1,6 +1,6 @@
 using Microsoft.AspNetCore.Mvc.Testing;
 
-namespace Reindeer.Web.Tests
+namespace Reindeer.Web.Tests.Common
 {
     public class ReindeerWebApplicationFactory : WebApplicationFactory<Program>
     {
